Moving relevant ISendOnlyBus methods to extension methods
